+++ packages/crdt_lf/CHANGELOG.md
@@ -4,6 +4,7 @@
 - chore: setup .github/workflows and update coverage links [33](https://github.com/MattiaPispisa/crdt/issues/33)
 - chore: update readme with recommended approach for complex handler types
 - chore: update topological sort implementation [3](https://github.com/MattiaPispisa/crdt/issues/3)
+- chore: added benchmarks
 
 ### Fixed
 - Fix readme reference links